# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "KnowledgeGap": {
        "properties": {
          "examples": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeGapExample"
            },
            "title": "Examples",
            "type": "array"
          },
          "first_seen_at": {
            "format": "date-time",
            "title": "First Seen At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_seen_at": {
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          },
          "occurrences": {
            "title": "Occurrences",
            "type": "integer"
          },
          "status": {
            "enum": [
              "open",
              "answered",
              "dismissed"
            ],
            "title": "Status",
            "type": "string"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "id",
          "topic",
          "status",
          "occurrences",
          "first_seen_at",
          "last_seen_at",
          "examples"
        ],
        "title": "KnowledgeGap",
        "type": "object"
      },
      "KnowledgeGapExample": {
        "properties": {
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "message_id": {
            "format": "uuid",
            "title": "Message Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "message_id",
          "conversation_id",
          "text",
          "occurred_at"
        ],
        "title": "KnowledgeGapExample",
        "type": "object"
      },
      "KnowledgeGapList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeGap"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "KnowledgeGapList",
        "type": "object"
      },
      "KnowledgeSource": {
        "properties": {
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "char_count": {
            "title": "Char Count",
            "type": "integer"
          },
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "file_asset_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Asset Id"
          },
          "file_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Name"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_ingested_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Ingested At"
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "status": {
            "enum": [
              "pending",
              "processing",
              "ready",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "enum": [
              "faq",
              "text",
              "url",
              "file"
            ],
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "type",
          "title",
          "status",
          "version",
          "char_count",
          "chunk_count",
          "created_at",
          "updated_at"
        ],
        "title": "KnowledgeSource",
        "type": "object"
      },
      "KnowledgeSourceCreate": {
        "additionalProperties": false,
        "description": "POST \u2026/knowledge-sources. FAQ: question + body (title = question). Note: title + body.\nWeb page: url (SEC-09). File: file_asset_id of an uploaded raw asset (PDF, DOCX, TXT, MD up to\n10 MB). ``gap_id`` answers a knowledge gap (FR-KB-06).",
        "properties": {
          "body": {
            "anyOf": [
              {
                "maxLength": 20000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "file_asset_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Asset Id"
          },
          "gap_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gap Id"
          },
          "question": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "type": {
            "enum": [
              "faq",
              "text",
              "url",
              "file"
            ],
            "title": "Type",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "type"
        ],
        "title": "KnowledgeSourceCreate",
        "type": "object"
      },
      "KnowledgeSourceList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeSource"
            },
            "title": "Items",
            "type": "array"
          },
          "usage": {
            "$ref": "#/components/schemas/KnowledgeUsage"
          }
        },
        "required": [
          "items",
          "usage"
        ],
        "title": "KnowledgeSourceList",
        "type": "object"
      },
      "KnowledgeSourcePatch": {
        "additionalProperties": false,
        "description": "PATCH \u2026/knowledge-sources/{id}: edits re-ingest the source (version + 1).",
        "properties": {
          "body": {
            "anyOf": [
              {
                "maxLength": 20000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "question": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "reingest": {
            "default": false,
            "title": "Reingest",
            "type": "boolean"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "url": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "title": "KnowledgeSourcePatch",
        "type": "object"
      },
      "KnowledgeTest": {
        "additionalProperties": false,
        "properties": {
          "question": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "KnowledgeTest",
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "items": {
              "properties": {
                "field": {
                  "title": "Field",
                  "type": "string"
                },
                "message": {
                  "title": "Message",
                  "type": "string"
                }
              },
              "required": [
                "field",
                "message"
              ],
              "type": "object"
            },
            "title": "Errors",
            "type": "array"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "WorkspaceOut": {
        "properties": {
          "automation_disclosure": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Automation Disclosure"
          },
          "checklist_dismissed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checklist Dismissed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "plan": {
            "enum": [
              "free",
              "pro",
              "max"
            ],
            "title": "Plan",
            "type": "string"
          },
          "reply_language": {
            "title": "Reply Language",
            "type": "string"
          },
          "role": {
            "enum": [
              "owner",
              "admin",
              "agent"
            ],
            "title": "Role",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "deleting"
            ],
            "title": "Status",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "timezone",
          "reply_language",
          "status",
          "role",
          "plan",
          "created_at"
        ],
        "title": "WorkspaceOut",
        "type": "object"
      },
      "WorkspacePatch": {
        "additionalProperties": false,
        "properties": {
          "automation_disclosure": {
            "anyOf": [
              {
                "maxLength": 60,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Automation Disclosure"
          },
          "checklist_dismissed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checklist Dismissed"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "reply_language": {
            "anyOf": [
              {
                "maxLength": 35,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reply Language"
          },
          "slug": {
            "anyOf": [
              {
                "maxLength": 48,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "timezone": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "title": "WorkspacePatch",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Social Hood API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/w/{wid}": {
      "get": {
        "description": "Also remembers this workspace as the user's last one, for /app (F-02).",
        "operationId": "get_workspace",
        "parameters": [
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Wid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (application/problem+json)"
          }
        },
        "summary": "Get Workspace",
        "tags": [
          "workspaces"
        ]
      },
      "patch": {
        "operationId": "update_workspace",
        "parameters": [
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Wid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspacePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (application/problem+json)"
          }
        },
        "summary": "Update Workspace",
        "tags": [
          "workspaces"
        ]
      }
    },
...
    "/v1/w/{wid}/knowledge-gaps": {
      "get": {
        "description": "Open gaps from the last 30 days, most asked first (TR-AI-12).",
        "operationId": "list_knowledge_gaps",
        "parameters": [
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Wid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "open",
              "pattern": "^(open|answered|dismissed)$",
              "title": "Status",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeGapList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (application/problem+json)"
          }
        },
        "summary": "List Knowledge Gaps",
        "tags": [
          "knowledge"
        ],
        "x-pending": "T5.10"
      }
    },
    "/v1/w/{wid}/knowledge-gaps/{gap_id}/dismiss": {
      "post": {
        "operationId": "dismiss_knowledge_gap",
        "parameters": [
          {
            "in": "path",
            "name": "gap_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Gap Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Wid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeGap"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (application/problem+json)"
          }
        },
        "summary": "Dismiss Knowledge Gap",
        "tags": [
          "knowledge"
        ],
        "x-pending": "T5.10"
      }
    },
    "/v1/w/{wid}/knowledge-sources": {
      "get": {
        "operationId": "list_knowledge_sources",
        "parameters": [
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Wid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSourceList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (application/problem+json)"
          }
        },
        "summary": "List Knowledge Sources",
        "tags": [
          "knowledge"
        ],
        "x-pending": "T5.3"
      },
      "post": {
        "description": "402 quota_exceeded over knowledge_characters; ingestion runs in the background.",
        "operationId": "create_knowledge_source",
        "parameters": [
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Wid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeSourceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (application/problem+json)"
          }
        },
        "summary": "Create Knowledge Source",
        "tags": [
          "knowledge"
        ],
        "x-pending": "T5.3"
      }
    },
    "/v1/w/{wid}/knowledge-sources/{source_id}": {
      "delete": {
        "operationId": "delete_knowledge_source",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Wid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (application/problem+json)"
          }
        },
        "summary": "Delete Knowledge Source",
        "tags": [
          "knowledge"
        ],
        "x-pending": "T5.3"
      },
      "get": {
        "operationId": "get_knowledge_source",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Wid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (application/problem+json)"
          }
        },
        "summary": "Get Knowledge Source",
        "tags": [
          "knowledge"
        ],
        "x-pending": "T5.3"
      },
      "patch": {
        "operationId": "update_knowledge_source",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Wid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeSourcePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (application/problem+json)"
          }
        },
        "summary": "Update Knowledge Source",
        "tags": [
          "knowledge"
        ],
        "x-pending": "T5.3"
      }
    },
    "/v1/w/{wid}/knowledge/test": {
      "post": {
        "description": "FR-KB-03: 1 credit; nothing is stored or sent.",
        "operationId": "test_knowledge",
        "parameters": [
          {
            "in": "path",
            "name": "wid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Wid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeTest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeTestResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Problem details (application/problem+json)"
          }
        },
        "summary": "Test Knowledge",
        "tags": [
          "knowledge"
        ],
        "x-pending": "T5.3"
      }
    },
...
  }
}